Refactor: rearranged the code that extends sdk functionality
The methods dealing with bank slip (issuing, registering and searching) and payout were moved to the class ExtendedPaymentWebService.
