+++ api/services/ExtendedPaymentWebService.ts
@@ -23,6 +23,22 @@ export class WithdrawModel extends BaseModel {
     }
 }
 
+export interface BoletoRequestSchema {
+    amount: string;
+    expiresAt: Date;
+}
+
+export class BoletoModel extends BaseModel {
+    transaction: Transaction;
+    recipients: Recipient[];
+    amount?: number;
+
+    constructor(data: Partial<BoletoModel> = {}) {
+        super(data);
+        Object.assign(this, data);
+    }
+}
+
 export default class ExtendedPaymentWebService extends PaymentWebService {
     
     constructor(options: PaymentWebServiceOptions) {
@@ -49,4 +65,50 @@ export default class ExtendedPaymentWebService extends PaymentWebService {
         return response.data;
       }
 
+      public async issueBankSlip(request: BoletoRequestSchema): Promise<BoletoModel> {
+        const { amount, expiresAt } = request;
+    
+        const url = `/payments/boleto`;
+        const body = { amount, expiresAt };
+        const signature = RequestUtil.sign(this.options.clientSecret, {
+          url,
+          method: "POST",
+          body: JSON.stringify(body)
+        });
+    
+        const response = await this.http.post(url, body, { headers: { ...signature } });
+    
+        if (!response || response.status !== 200) {
+          throw response;
+        }
+    
+        return new BoletoModel(response.data);
+      }
+
+      public async registerBankSlip(bankSlipId: string): Promise<BoletoModel> {
+    
+        const url = `/payments/boleto/register/${bankSlipId}`;
+        const signature = RequestUtil.sign(this.options.clientSecret, {
+          url,
+          method: "POST"
+        });
+    
+        const response = await this.http.post(url, { headers: { ...signature } });
+    
+        if (!response || response.status !== 200) {
+          throw response;
+        }
+    
+        return new BoletoModel(response.data);
+      }
+
+      public async findBankSlipById(bankSlipId: string): Promise<BoletoModel> {
+        const response = await this.http.get(`/payments/boleto/${bankSlipId}`);
+    
+        if (!response || response.status !== 200) {
+          throw response;
+        }
+    
+        return new BoletoModel(response.data);
+      }
 }
